Extend max values for brewing fuel and brew time

+++ library/item/item_content_registry/src/main/java/org/quiltmc/qsl/item/content/registry/api/ItemContentRegistries.java
@@ -72,6 +72,6 @@ public class ItemContentRegistries {
 			.builder(Registry.ITEM,
 					new Identifier(NAMESPACE, "brewing_fuel_time"),
 					Integer.class,
-					Codec.intRange(0, Byte.MAX_VALUE))
+					Codec.intRange(0, Integer.MAX_VALUE))
 			.build();
 }
